refactor: App.tsx 리팩토링
